riscv64: Trap handler divided in submodules. Add doc
