+++ Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+num_enum = {version = "0.5.1", default-features = false} 
 qemu-exit = "1.0.0"
 cfg-if = "0.1.10"
 riscv = "0.6.0"
