+++ src/arch/riscv64/mod.rs
@@ -1,14 +1,14 @@
 //! RISCV64 abstractions layer
 
-mod interrupts;
+pub mod exceptions;
+pub mod interrupts;
+pub mod trap;
 
 use super::*;
 use crate::kmain;
 use crate::println;
 use crate::utils::external_symbol_address;
 use cfg_if::cfg_if;
-
-use interrupts::InterruptsState;
 
 /// UART0 address
 pub static UART0: usize = 0x10000000;
